Refactor JSONArray and add comments
